Move Proxy to dedicated protocol directory
