--- docs/source/api/reference/conpot.protocols.misc.rst
@@ -1,22 +1,0 @@
-conpot.protocols.misc package
-=============================
-
-Submodules
-----------
-
-conpot.protocols.misc.ascii\_decoder module
--------------------------------------------
-
-.. automodule:: conpot.protocols.misc.ascii_decoder
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: conpot.protocols.misc
-   :members:
-   :undoc-members:
-   :show-inheritance:
